Fix command for getting project version
The problem was with the grep command trying to extract version from
the output of Maven. When Maven downloaded dependencies, it included
them in the result. Corrected script suppresses all output except of
the version number, so grep is not needed. Solution is based on
the blog post: https://blog.soebes.de/blog/2018/06/09/help-plugin/